+++ lua/obsidian/commands/workspace.lua
@@ -1,32 +1,35 @@
-local log = require "obsidian.log"
 local Workspace = require "obsidian.workspace"
 
 ---@param data CommandArgs
 return function(_, data)
   if not data.args or string.len(data.args) == 0 then
     local picker = Obsidian.picker
-    if not picker then
-      log.info("Current workspace: '%s' @ '%s'", Obsidian.workspace.name, Obsidian.workspace.path)
-      return
+    if picker then
+      ---@type obsidian.PickerEntry
+      local options = vim.tbl_map(function(ws)
+        return {
+          value = ws,
+          display = tostring(ws),
+          filename = tostring(ws.path),
+        }
+      end, Obsidian.workspaces)
+      picker:pick(options, {
+        prompt_title = "Obsidian Workspace",
+        callback = function(ws)
+          Workspace.switch(ws.name, { lock = true })
+        end,
+      })
+    else
+      vim.ui.select(Obsidian.workspaces, {
+        prompt = "Obsidian Workspace",
+        format_item = tostring,
+      }, function(ws)
+        if not ws then
+          return
+        end
+        Workspace.switch(ws.name, { lock = true })
+      end)
     end
-
-    local options = {}
-    for i, spec in ipairs(Obsidian.opts.workspaces) do
-      local workspace = Workspace.new_from_spec(spec)
-      if workspace == Obsidian.workspace then
-        options[#options + 1] = string.format("*[%d] %s @ '%s'", i, workspace.name, workspace.path)
-      else
-        options[#options + 1] = string.format("[%d] %s @ '%s'", i, workspace.name, workspace.path)
-      end
-    end
-
-    picker:pick(options, {
-      prompt_title = "Workspaces",
-      callback = function(workspace_str)
-        local idx = tonumber(string.match(workspace_str, "%*?%[(%d+)]"))
-        Workspace.switch(Obsidian.opts.workspaces[idx].name, { lock = true })
-      end,
-    })
   else
     Workspace.switch(data.args, { lock = true })
   end