+++ lua/obsidian/workspace.lua
@@ -1,22 +1,15 @@
 local Path = require "obsidian.path"
 local abc = require "obsidian.abc"
-local api = require "obsidian.api"
 local util = require "obsidian.util"
 local config = require "obsidian.config"
 local log = require "obsidian.log"
 
 ---@class obsidian.workspace.WorkspaceSpec
 ---
----@field path string|(fun(): string)
+---@field path string|(fun(): string)|obsidian.Path|(fun(): obsidian.Path)
 ---@field name string|?
 ---@field strict boolean|? If true, the workspace root will be fixed to 'path' instead of the vault root (if different).
 ---@field overrides table|obsidian.config.ClientOpts?
-
----@class obsidian.workspace.WorkspaceOpts
----
----@field name string|?
----@field strict boolean|? If true, the workspace root will be fixed to 'path' instead of the vault root (if different).
----@field overrides table|obsidian.config.ClientOpts|?
 
 --- Each workspace represents a working directory (usually an Obsidian vault) along with
 --- a set of configuration options specific to the workspace.
@@ -35,7 +28,10 @@ local log = require "obsidian.log"
 ---@field locked boolean|?
 local Workspace = abc.new_class {
   __tostring = function(self)
-    return string.format("Workspace(name='%s', path='%s', root='%s')", self.name, self.path, self.root)
+    if self.name == Obsidian.workspace.name then
+      return string.format("*[%s] @ '%s'", self.name, self.path)
+    end
+    return string.format("[%s] @ '%s'", self.name, self.path)
   end,
   __eq = function(a, b)
     local a_fields = a:as_tbl()
@@ -72,19 +68,29 @@ end
 
 --- Create a new 'Workspace' object. This assumes the workspace already exists on the filesystem.
 ---
----@param path string|obsidian.Path Workspace path.
----@param opts obsidian.workspace.WorkspaceOpts|?
+---@param spec obsidian.workspace.WorkspaceSpec|?
 ---
 ---@return obsidian.Workspace
-Workspace.new = function(path, opts)
-  opts = opts and opts or {}
+Workspace.new = function(spec)
+  spec = spec and spec or {}
+
+  local path
+
+  if type(spec.path) == "function" then
+    path = spec.path()
+  else
+    path = spec.path
+  end
+
+  ---@cast path -function
+  path = vim.fs.normalize(tostring(path))
 
   local self = Workspace.init()
   self.path = Path.new(path):resolve { strict = true }
-  self.name = assert(opts.name or self.path.name)
-  self.overrides = opts.overrides
+  self.name = assert(spec.name or self.path.name)
+  self.overrides = spec.overrides
 
-  if opts.strict then
+  if spec.strict then
     self.root = self.path
   else
     local vault_root = find_vault_root(self.path)
@@ -96,29 +102,6 @@ Workspace.new = function(path, opts)
   end
 
   return self
-end
-
---- Initialize a new 'Workspace' object from a workspace spec.
----
----@param spec obsidian.workspace.WorkspaceSpec
----
----@return obsidian.Workspace
-Workspace.new_from_spec = function(spec)
-  ---@type string|obsidian.Path
-  local path
-  if type(spec.path) == "function" then
-    path = spec.path()
-  else
-    ---@diagnostic disable-next-line: cast-local-type
-    path = spec.path
-  end
-
-  ---@diagnostic disable-next-line: param-type-mismatch
-  return Workspace.new(path, {
-    name = spec.name,
-    strict = spec.strict,
-    overrides = spec.overrides,
-  })
 end
 
 --- Lock the workspace.
@@ -149,103 +132,49 @@ Workspace.get_workspace_for_dir = function(cur_dir, workspaces)
   end
 
   for _, spec in ipairs(workspaces) do
-    local w = Workspace.new_from_spec(spec)
+    local w = Workspace.new(spec)
     if w.path == cur_dir or w.path:is_parent_of(cur_dir) then
       return w
     end
   end
 end
 
---- Get the workspace corresponding to the current working directory (or a parent of), if there
---- is one.
+--- 1. Set Obsidian.workspace, Obsidian.dir, and opts
+--- 2. Make sure all the directories exists
+--- 3. fire callbacks and exec autocmd event
 ---
----@param workspaces obsidian.workspace.WorkspaceSpec[]
----
----@return obsidian.Workspace|?
-Workspace.get_workspace_for_cwd = function(workspaces)
-  local cwd = assert(vim.fn.getcwd())
-  return Workspace.get_workspace_for_dir(cwd, workspaces)
-end
-
---- Returns the default workspace.
----
----@param workspaces obsidian.workspace.WorkspaceSpec[]
----
----@return obsidian.Workspace|nil
-Workspace.get_default_workspace = function(workspaces)
-  if not vim.tbl_isempty(workspaces) then
-    return Workspace.new_from_spec(workspaces[1])
-  else
-    return nil
-  end
-end
-
---- Resolves current workspace from the client config.
----
----@param opts obsidian.config.ClientOpts
----
----@return obsidian.Workspace|?
-Workspace.get_from_opts = function(opts)
-  local current_workspace = Workspace.get_workspace_for_cwd(opts.workspaces)
-
-  if not current_workspace then
-    current_workspace = Workspace.get_default_workspace(opts.workspaces)
-  end
-
-  return current_workspace
-end
-
---- Get the normalize opts for a given workspace.
----
----@param workspace obsidian.Workspace|?
----
----@return obsidian.config.ClientOpts
-Workspace.normalize_opts = function(workspace)
-  if workspace then
-    return config.normalize(workspace.overrides and workspace.overrides or {}, Obsidian._opts)
-  else
-    return Obsidian.opts
-  end
-end
-
 ---@param workspace obsidian.Workspace
 ---@param opts { lock: boolean|? }|?
 Workspace.set = function(workspace, opts)
   opts = opts and opts or {}
 
   local dir = workspace.root
-  local options = Workspace.normalize_opts(workspace) -- TODO: test
+  local options = config.normalize(workspace.overrides, Obsidian._opts)
 
   Obsidian.workspace = workspace
   Obsidian.dir = dir
   Obsidian.opts = options
 
   -- Ensure directories exist.
-  dir:mkdir { parents = true, exists_ok = true }
+  dir:mkdir { parents = true }
 
-  if options.notes_subdir ~= nil then
-    local notes_subdir = dir / Obsidian.opts.notes_subdir
-    notes_subdir:mkdir { parents = true, exists_ok = true }
+  if options.notes_subdir then
+    (dir / options.notes_subdir):mkdir { parents = true }
   end
 
-  if Obsidian.opts.daily_notes.folder ~= nil then
-    local daily_notes_subdir = Obsidian.dir / Obsidian.opts.daily_notes.folder
-    daily_notes_subdir:mkdir { parents = true, exists_ok = true }
+  if options.templates.folder then
+    (dir / options.templates.folder):mkdir { parents = true }
   end
 
-  -- Setup UI add-ons.
-  local has_no_renderer = not (api.get_plugin_info "render-markdown.nvim" or api.get_plugin_info "markview.nvim")
-  if has_no_renderer and Obsidian.opts.ui.enable then
-    require("obsidian.ui").setup(Obsidian.workspace, Obsidian.opts.ui)
+  if options.daily_notes.folder then
+    (dir / options.daily_notes.folder):mkdir { parents = true }
   end
 
   if opts.lock then
     Obsidian.workspace:lock()
   end
 
-  Obsidian.picker = require("obsidian.pickers").get(Obsidian.opts.picker.name)
-
-  util.fire_callback("post_set_workspace", Obsidian.opts.callbacks.post_set_workspace, workspace)
+  util.fire_callback("post_set_workspace", options.callbacks.post_set_workspace, workspace)
 
   vim.api.nvim_exec_autocmds("User", {
     pattern = "ObsidianWorkpspaceSet",
@@ -253,7 +182,7 @@ Workspace.set = function(workspace, opts)
   })
 end
 
----@param workspace obsidian.Workspace|string The workspace object or the name of an existing workspace.
+---@param workspace string name of workspace
 ---@param opts { lock: boolean|? }|?
 Workspace.switch = function(workspace, opts)
   opts = opts and opts or {}
@@ -265,7 +194,7 @@ Workspace.switch = function(workspace, opts)
 
   for _, ws in ipairs(Obsidian.opts.workspaces) do
     if ws.name == workspace then
-      return Workspace.set(Workspace.new_from_spec(ws), opts)
+      return Workspace.set(Workspace.new(ws), opts)
     end
   end