+++ tests/test_workspace.lua
@@ -7,7 +7,10 @@ local T = new_set()
 T["should be able to initialize a workspace"] = function()
   local tmpdir = Path.temp()
   tmpdir:mkdir()
-  local ws = workspace.new(tmpdir, { name = "test_workspace" })
+  local ws = workspace.new {
+    path = tmpdir,
+    name = "test_workspace",
+  }
   eq("test_workspace", ws.name)
   eq(true, tmpdir:resolve() == ws.path)
   tmpdir:rmdir()