+++ CHANGELOG.md
@@ -75,6 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed tag completion for blink.cmp and improved frontmatter tag handling
 - Fixed to allow numbers in note IDs.
 - Fixed ignore subdirectories specified in `daily_notes.date_format`
+- Fixed not make sure template folder exists.
+- Refactored workspace module for a better api.
 
 ## [v3.12.0](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.12.0) - 2025-06-05